requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
certifi==2025.7.14
colorama==0.4.6
distro==1.9.0
Django==5.2.4
faiss-cpu==1.11.0.post1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==2.3.2
openai==1.97.1
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
